group26K/backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.0
django-environ>=0.11
psycopg[binary]>=3.1
djangorestframework-gis>=1.0
wagtail==7.0


